fully working scanner script
The previous version of the script only contained a function to run on
each file. This pulls over the commands I was running to get the files
and tidies up the script a bit.

It'll now build an initial set of "create" events from a directory,
ignoring files that it's seen before.