add db audit script
Since everything right now is very dynamic, it would be really easy for
typos to creep in and corrupt the database. The first thing I noticed
was that the event type is one of a few (three right now) distinct
values, and if some script uses a bad value, like "created" instead of
"create", then everything falls over.

This attempts to address the issue by adding a script that can audit the
db and be sure that everything looks as expected. For now, it only looks
at event type. Add other checks as the need becomes apparent.