refactor scripts
Not quite a true refactoring, but close. All I did was relocate some
functions and name them appropriately, plus making the scanner script do
both the "create/update" and "delete" scans all in one go. That's the
change that isn't a true refactoring. Otherwise, everything is the same,
just reorganized.