add java version
It was so long since I touched this last that I forgot where I was and
just started tinkering with some java code to build a database of a
photo directory. I ended up basically recreating what the bash script
does, but I'm pretty sure it's at least twice as performant because it
handles db stuff more efficiently. Also, tests!

This is a pretty early working version, and I'm actually currently
comparing a db generated by the shell version with one generated by the
java version to see if they come up with the same results.

Still todo:

- put in real logging so i can control log levels

- maybe put in a real metrics system