more testing
Adds another IT to test file changes on the other side from what's
already tested, and beefs up the baseline test a bit. I had tried
filtering recently changed files in the catalog builder to avoid
in-progress syncing, but that obviously messed with the tests. As I
tried to sort that out, I found that the age check wasn't even working
right, so I've left it out for now but kept the test improvements.