extract file system scanner
The current java implementation is only half of the picture. It scans
the file system and updates the database. It also has to check the
database to see if files have been removed. This moves the file scanning
into its own class, since things are too complex already. Now I should
have room to add in that second part.